# swagger/swagger.yaml
openapi: "3.0.1"
info:
  description: "This is a swagger for Medical Virtual Center (MVC)"
  version: "1.0.0"
  title: "Medical Virtual Center"
servers:
  - url: http://localhost:84
    description: develop
...
paths:
...
  /mvc/schedule/add-note/{scheduleId}:
    put:
      tags:
        - "Schedule"
      summary: "add note in schedule"
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleAddNoteInput'
      responses:
        200:
          description: OK
          content:
            application/json:
              examples:
                string:
                  value: "Se ha añadido la nota correctamente"
        401:
          description: Unauthorized
          content:
            application/json:
              examples:
                json:
                  value: { "error": "Invalid credentials." }
      security:
        - token: [ ]
...
components:
  schemas:
...
    ScheduleAddNoteInput:
      type: object
      properties:
        date:
          type: string
          nullable: false
        text:
          type: string
          nullable: false
...
  securitySchemes:
    token:
      type: apiKey
      name: X-AUTH-TOKEN
      description: User api key for third-party
      in: header